% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020.R
\name{calc_db_cutoff}
\alias{calc_db_cutoff}
\title{Calculate cutoff}
\usage{
calc_db_cutoff(income, db_cutoff1, inflation)
}
\arguments{
\item{income}{output from \code{income_projection}}

\item{db_cutoff1}{income cutoff year 1}

\item{inflation}{inflation}
}
\value{
Tibble
}
\description{
Calculate cutoff
}
