% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{calc_db_dc}
\alias{calc_db_dc}
\title{Calculate pension with combined DB and DC components}
\usage{
calc_db_dc(
  ret,
  annuity,
  income,
  employee_cont,
  employer_cont,
  prop_salary,
  db_cutoff,
  incr,
  mult
)
}
\arguments{
\item{ret}{Investment returns per year}

\item{annuity}{output from \code{annuity_rates}}

\item{income}{output from \code{income_projection}}

\item{employee_cont}{Percentage of salary contributed by employee}

\item{employer_cont}{Percentage of salary contributed by employer}

\item{prop_salary}{Proportion of salary contributing to DB}

\item{db_cutoff}{Salary cutoff from DB to DC}

\item{incr}{DB Parameter}

\item{mult}{DB Parameter}
}
\value{
Tibble
}
\description{
Calculate pension with combined DB and DC components
}
