% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020.R
\name{calc_tax_thresholds}
\alias{calc_tax_thresholds}
\title{Tax thresholds}
\usage{
calc_tax_thresholds(income, inflation)
}
\arguments{
\item{income}{output from \code{income_projection}}

\item{inflation}{inflation}
}
\value{

}
\description{
Tax thresholds
}
