% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{constantCostFunction}
\alias{constantCostFunction}
\title{Cost function for constant amount}
\usage{
constantCostFunction(savedPerYear, targetTotal, years, prudence, fund)
}
\arguments{
\item{savedPerYear}{Total savings target}

\item{years}{Number of years to calculate}

\item{prudence}{Parameter for \code{investment_returns}, 50 or 65}

\item{fund}{Parameter for \code{investment_returns}, "USS", "Growth fund", "Moderate growth fund", "Cautious growth fund", or "Cash fund"}
}
\value{
Dollars needed to save per year.
}
\description{
This function is the cost function used for determining how much is needed 
to be saved as a constant amount per year.
}
