% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020.R
\name{contribution_rates}
\alias{contribution_rates}
\title{Contribution rates}
\usage{
contribution_rates(nyears, contribution_model = "USS Trustee current plans")
}
\arguments{
\item{nyears}{how many years to calculate}

\item{contribution_model}{Default ="USS Trustee current plans"}
}
\value{
Tibble
}
\description{
<full description>
}
