% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020.R
\name{contributions_model}
\alias{contributions_model}
\title{Contributions over all models}
\usage{
contributions_model(
  income,
  db_cutoff1 = 58589.7,
  inflation = 0.02,
  lower_rate = 0.2,
  higher_rate = 0.4,
  additional_rate = 0.45
)
}
\arguments{
\item{income}{output from \code{income_projection}}

\item{db_cutoff1}{income cutoff year 1 default=58589.7}

\item{inflation}{default=0.02}

\item{lower_rate}{default= 0.2}

\item{higher_rate}{default =0.4}

\item{additional_rate}{default=0.45}
}
\value{
Tibble
}
\description{
Contributions over all models
}
