% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{income_projection}
\alias{income_projection}
\title{Income calculator}
\usage{
income_projection(
  start,
  increase,
  increases_per_year = 1,
  years,
  upper_limit = 1e+06
)
}
\arguments{
\item{start}{Starting income}

\item{increase}{Percentage increase}

\item{increases_per_year}{Number of increases per year. Default 1}

\item{years}{Number of years to calculate}

\item{upper_limit}{Maximum income. Default 1000000}
}
\value{
vector of incomes
}
\description{
Given a starting income and an annual growth, will return projected income for given number of years. Upper limit can be set also.
}
