% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{investment_returns}
\alias{investment_returns}
\title{Calculation for investment returns}
\usage{
investment_returns()
}
\value{
Data frame for different schemes and over time
}
\description{
Calculation for investment returns
}
