% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020_valuation.R
\name{investment_returns_2020}
\alias{investment_returns_2020}
\title{Investment returns for the 2020 valuation}
\usage{
investment_returns_2020(years)
}
\value{
Data frame
}
\description{
Investment returns for the 2020 valuation
}
