% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{investment_savings_total}
\alias{investment_savings_total}
\title{Savings calculator}
\usage{
investment_savings_total(savedPerYear, years, prudence, fund)
}
\arguments{
\item{savedPerYear}{Constant amount saved per year}

\item{years}{Number of years to calculate}

\item{prudence}{Parameter for \code{investment_returns}, 50 or 65}

\item{fund}{Parameter for \code{investment_returns}, "USS", "Growth fund", "Moderate growth fund", "Cautious growth fund", or "Cash fund"}
}
\value{
Dollars needed to save per year.
}
\description{
Given constant amount saved per year calculate total saved at end of years
}
