% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{pension_calculation}
\alias{pension_calculation}
\title{Calculate pensions for different schemes}
\usage{
pension_calculation(income, annuity, prudence, fund)
}
\arguments{
\item{income}{output from \code{income_projection}}

\item{annuity}{output from \code{annuity_rates}}

\item{prudence}{Parameter for \code{investment_returns}, 50 or 65}

\item{fund}{Parameter for \code{investment_returns}, "USS", "Growth fund", "Moderate growth fund", "Cautious growth fund", or "Cash fund"}
}
\value{
Data frame of years and pension pots and annual benefits
}
\description{
Calculates total pot and annual benefits for DB, DC and TPS pensions
}
