% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020_valuation.R
\name{pension_calculation_2020}
\alias{pension_calculation_2020}
\title{Pension scenarios 2020}
\usage{
pension_calculation_2020(income, annuity, scenario)
}
\arguments{
\item{income}{Output from income_projection()}

\item{annuity}{Output from annuity_rates()}

\item{scenario}{Investment scenario for the DC component of the current scheme ("Scenario 1", "Scenario 2a", "Scenario 2b", "Scenario 3a", "Scenario 3b")}
}
\value{
List of data frames for each scenario
}
\description{
Pension scenarios 2020
}
