% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020_valuation.R
\name{pension_calculation_2020_summary}
\alias{pension_calculation_2020_summary}
\title{Summarise the output from pension_calculation_2020()}
\usage{
pension_calculation_2020_summary(pension)
}
\arguments{
\item{pension}{Output from pension_calculation_2020()}
}
\value{
data frame
}
\description{
Summarise the output from pension_calculation_2020()
}
