% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{pension_summary}
\alias{pension_summary}
\title{Pension summary}
\usage{
pension_summary(benefits, dob)
}
\arguments{
\item{benefits}{Output from \code{pension_calculation}}

\item{dob}{e.g. "06/09/1984" for 6 september 1984}
}
\value{
relevant row from output of \code{pension_calculation}
}
\description{
Calculate various differences
}
