% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{percentage_savings_total}
\alias{percentage_savings_total}
\title{Savings calculator}
\usage{
percentage_savings_total(savedPerYear, income, prudence, fund)
}
\arguments{
\item{savedPerYear}{Percentage saved per year}

\item{income}{vector of projected income values for each year}

\item{prudence}{Parameter for \code{investment_returns}, 50 or 65}

\item{fund}{Parameter for \code{investment_returns}, "USS", "Growth fund", "Moderate growth fund", "Cautious growth fund", or "Cash fund"}
}
\value{
Dollars needed to save per year.
}
\description{
Given percent of salary saved per year calculate total saved at end of years
}
