% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2020_valuation.R
\name{plot_pension}
\alias{plot_pension}
\title{Plot pension}
\usage{
plot_pension(pension, years, column = "total_pot")
}
\arguments{
\item{pension}{Output from pension_calculation_2020()}

\item{column}{Defailt="total_pot"}
}
\value{
plot
}
\description{
Plot pension
}
