% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retirement_date}
\alias{retirement_date}
\title{Retirement date calculator}
\usage{
retirement_date(dob)
}
\arguments{
\item{dob}{e.g. "06/09/1984" for 6 september 1984}
}
\value{
date object
}
\description{
Just adds 68 years. Needs to be made more accurate based on birth year
}
