% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{savings_required_percentage}
\alias{savings_required_percentage}
\title{Savings calculator}
\usage{
savings_required_percentage(amount, income, prudence, fund)
}
\arguments{
\item{amount}{Total savings target}

\item{income}{Projected income each year}

\item{prudence}{Parameter for \code{investment_returns}, 50 or 65}

\item{fund}{Parameter for \code{investment_returns}, "USS", "Growth fund", "Moderate growth fund", "Cautious growth fund", or "Cash fund"}
}
\value{
Dollars needed to save per year.
}
\description{
Given a total savings target calculalate how much per 
year needed to achieve target savings total
}
