% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2018.R
\name{years_left}
\alias{years_left}
\title{Years of work left until retirement}
\usage{
years_left(dob)
}
\arguments{
\item{dob}{e.g. "06/09/1984" for 6 september 1984}
}
\value{
Numeric value
}
\description{
Years of work left until retirement
}
