% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annuity_rates}
\alias{annuity_rates}
\title{Calculate annuity rates}
\usage{
annuity_rates(sex, type, years, le_increase = 0.015)
}
\arguments{
\item{sex}{"men" or "women"}

\item{type}{"single" or "joint"}

\item{years}{number of years to calculate}

\item{le_increase}{Expected increase in life expectancy per year. Default 0.015}
}
\value{
vector of annuities
}
\description{
Different rates for sex, joint vs single, and expected annual life expectancy increase
}
